feat: add lambda to invoke reindex api (#23)
In order to periodically reindex the Momento docs and blogs, we add a
lambda that invokes the reindex API.

The reindex API runs a background task asynchronously that fetches the
Momento docs and blogs, chunks, embeds, and stores in MVI.

The lambda assumes two environment variables are set:
ROBOMO_API_ENDPOINT and ROBOMO_INDEX_NAME. The former should be the
endpoint of the server with the /reindex path. The latter is the name
of the index to use to store the indexed data.