chore: comment out main stack for now (#27)
* chore: comment out main stack for now

We need the DNS stacks to get deployed to the accounts so that
we can set up delegation. We can't deploy the second stack yet
because it has a dependency on the DNS hosted zone. This commit
comments out the main stack for now, and we can uncomment it
once the DNS is up.